feature: New 'plain' actor type
* refactor: Introduced new intermediary class 'AmbersteelBaseCharacterActor' and moved some methods over to it, extracted from 'AmbersteelBaseActor'

* refactor: Adjusted templates to accomodated new actor type.

* refactor: Renamed 'actor-character-sheet' -> 'actor-sheet'.

* fix: Chat messages originating in actors no longer throw an error that 'getChatViewModel' is undefined.
