+++ templates/actor/actor-chat-message.hbs
@@ -15,47 +15,59 @@ cssClass: {undefined | String}
       {{viewModel.actor.name}}
     </h2>
   </div>
-  {{!-- Personals --}}
-  {{!-- Species --}}
-  <div class="flex flex-row sheet-block margin-b-md">
-    <div class="width-min-xl fancy-font">
-      {{localize "ambersteel.personals.species"}}
+  {{#if (eq viewModel.isPlain false)}}
+    {{!-- Personals --}}
+    {{!-- Species --}}
+    <div class="flex flex-row sheet-block margin-b-md">
+      <div class="width-min-xl fancy-font">
+        {{localize "ambersteel.personals.species"}}
+      </div>
+      <div class="flex-grow">
+        {{viewModel.actor.data.data.person.species}}
+      </div>
     </div>
-    <div class="flex-grow">
-      {{viewModel.actor.data.data.person.species}}
+    <div class="flex flex-row sheet-block margin-b-md">
+      <div class="width-min-xl fancy-font">
+        {{localize "ambersteel.personals.culture"}}
+      </div>
+      <div class="flex-grow">
+        {{viewModel.actor.data.data.person.culture}}
+      </div>
     </div>
-  </div>
-  <div class="flex flex-row sheet-block margin-b-md">
-    <div class="width-min-xl fancy-font">
-      {{localize "ambersteel.personals.culture"}}
+    <div class="flex flex-row sheet-block margin-b-md">
+      <div class="width-min-xl fancy-font">
+        {{localize "ambersteel.personals.sex"}}
+      </div>
+      <div class="flex-grow">
+        {{viewModel.actor.data.data.person.sex}}
+      </div>
     </div>
-    <div class="flex-grow">
-      {{viewModel.actor.data.data.person.culture}}
+    <div class="flex flex-row sheet-block margin-b-md">
+      <div class="width-min-xl fancy-font">
+        {{localize "ambersteel.personals.age"}}
+      </div>
+      <div class="flex-grow">
+        {{viewModel.actor.data.data.person.age}}
+      </div>
     </div>
-  </div>
-  <div class="flex flex-row sheet-block margin-b-md">
-    <div class="width-min-xl fancy-font">
-      {{localize "ambersteel.personals.sex"}}
+    {{!-- Biography --}}
+    <div class="sheet-block">
+      <h3 class="width-min-xl fancy-font">
+        {{localize "ambersteel.labels.biography"}}
+      </h3>
+      <div class="scrollable-height-md">
+        {{{viewModel.renderedBiography}}}
+      </div>
     </div>
-    <div class="flex-grow">
-      {{viewModel.actor.data.data.person.sex}}
+  {{else}}
+    {{!-- Description --}}
+    <div class="sheet-block">
+      <h3 class="width-min-xl fancy-font">
+        {{localize "ambersteel.labels.description"}}
+      </h3>
+      <div class="scrollable-height-md">
+        {{{viewModel.renderedDescription}}}
+      </div>
     </div>
-  </div>
-  <div class="flex flex-row sheet-block margin-b-md">
-    <div class="width-min-xl fancy-font">
-      {{localize "ambersteel.personals.age"}}
-    </div>
-    <div class="flex-grow">
-      {{viewModel.actor.data.data.person.age}}
-    </div>
-  </div>
-  {{!-- Biography --}}
-  <div class="sheet-block">
-    <h3 class="width-min-xl fancy-font">
-      {{localize "ambersteel.labels.biography"}}
-    </h3>
-    <div class="scrollable-height-md">
-      {{{viewModel.renderedBiography}}}
-    </div>
-  </div>
+  {{/if}}
 </section>
